Test: Add a new test suite for run key migration indepotency
